app/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
bandit==1.7.5
billiard==4.1.0
black
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
Django==4.2.6
djangorestframework==3.14.0
dparse==0.6.3
drf-yasg==1.21.7
exceptiongroup==1.1.3
flake8==6.1.0
flower==2.0.1
gitdb==4.0.10
GitPython==3.1.37
humanize==4.8.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
kombu==5.3.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pbr==5.11.1
pep8-naming==0.13.3
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.1.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rich==13.6.0
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safety
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
stevedore==5.1.0
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
